pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ch340_nrf"
version = "0.1.2"
authors = [{ name = "drakeerv", email = "drakeerv@outlook.com" }]
description = "Allows you to use the ch340 nrf usb adapter with python. This project is currently in developement."
readme = "README.md"
requires-python = ">=3.0"
classifiers = [
  "Programming Language :: Python :: 3",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: POSIX",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS :: MacOS X",
  "Development Status :: 3 - Alpha",
  "Topic :: Communications",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Terminals :: Serial",
]
dependencies = ["pyserial"]

[project.urls]
"Homepage" = "https://github.com/drakeerv/ch340_nrf"
"Bug Tracker" = "https://github.com/drakeerv/ch340_nrf/issues"
